+++ spec/query_builder_spec.rb
@@ -30,10 +30,15 @@ module OData
         builder.top(5)
         builder.query.should eq "Categories(1)/$links/Products?$top=5"
       end
-      it "should throw an execption if count was already called on the builder" do
+      it "should throw an exception if count was already called on the builder" do
         builder = QueryBuilder.new "Categories(1)"
         builder.count
         lambda { builder.links("Products") }.should raise_error(OData::NotSupportedError, "You cannot call both the `links` method and the `count` method in the same query.")
+      end
+      it "should throw an exception if select was already called on the builder" do
+        builder = QueryBuilder.new "Products"
+        builder.select "Price"
+        lambda { builder.links("Products") }.should raise_error(OData::NotSupportedError, "You cannot call both the `links` method and the `select` method in the same query.")
       end
     end
 
@@ -53,10 +58,63 @@ module OData
         builder.count
         builder.query.should eq "Products/$count?$filter=Name+eq+%27Widget+1%27"
       end
-      it "should throw an execption if links was already called on the builder" do
+      it "should throw an exception if links was already called on the builder" do
         builder = QueryBuilder.new "Categories(1)"
         builder.links("Products")
         lambda { builder.count }.should raise_error(OData::NotSupportedError, "You cannot call both the `links` method and the `count` method in the same query.")
+      end
+      it "should throw an exception if select was already called on the builder" do
+        builder = QueryBuilder.new "Products"
+        builder.select("Price")
+        lambda { builder.count }.should raise_error(OData::NotSupportedError, "You cannot call both the `select` method and the `count` method in the same query.")
+      end
+    end
+
+    describe "#select" do
+      it "should accept the select method" do
+        builder = QueryBuilder.new "Products"
+        lambda { builder.select }.should_not raise_error
+      end
+      it "should properly handle the select method" do
+        builder = QueryBuilder.new "Products"
+        builder.select "Price", "Rating"
+        builder.query.should eq "Products?$select=Price,Rating"
+      end
+      it "shouldn't add duplicate properties" do
+        builder = QueryBuilder.new "Products"
+        builder.select "Price", "Rating", "Price"
+        builder.query.should eq "Products?$select=Price,Rating"
+      end
+      it "should properly handle the select method with additional operators" do
+        builder = QueryBuilder.new "Products"
+        builder.filter("Name eq 'Widget 1'")
+        builder.select("Price", "Rating")
+        builder.query.should eq "Products?$select=Price,Rating&$filter=Name+eq+%27Widget+1%27"
+      end
+      it "should throw an exception if links was already called on the builder" do
+        builder = QueryBuilder.new "Categories(1)"
+        builder.links("Products")
+        lambda { builder.select("Price") }.should raise_error(OData::NotSupportedError, "You cannot call both the `links` method and the `select` method in the same query.")
+      end
+      it "should throw an exception if count was already called on the builder" do
+        builder = QueryBuilder.new "Products"
+        builder.count
+        lambda { builder.select("Price") }.should raise_error(OData::NotSupportedError, "You cannot call both the `count` method and the `select` method in the same query.")
+      end
+      it "should return itself" do
+        builder = QueryBuilder.new "Products"
+        result = builder.select("Price", "Rating")
+        result.should be_a(QueryBuilder)
+      end
+      it "should automatically add $expand for nested selects" do
+        builder = QueryBuilder.new "Categories"
+        result = builder.select "Name", "Products/Name"
+        result.query.should eq "Categories?$select=Name,Products/Name&$expand=Products"
+      end
+      it "should automatically add $expand for deeply nested selects" do
+        builder = QueryBuilder.new "Products"
+        result = builder.select "ProductName", "Order_Details/OrderID", "Order_Details/Order/Customer/CompanyName"
+        result.query.should eq "Products?$select=ProductName,Order_Details/OrderID,Order_Details/Order/Customer/CompanyName&$expand=Order_Details,Order_Details/Order/Customer"
       end
     end